リファクタリング：ツイートカードボタンを分割
