+++ src/features/tweet/views/TweetCard.tsx
@@ -3,7 +3,8 @@ import { MainAvatar } from "components/avatar/MainAvatar";
 import { MainButton } from "components/button/MainButton";
 import { BooleanLink } from "components/link/BooleanLink";
 import { Tweet } from "features/tweet/tweetTypes";
-import { TweetCardButton, CommentButton } from "features/tweet/views/TweetCardButton";
+import { TweetCardButton } from "features/tweet/views/TweetCardButton/TweetCardButton";
+import { CommentButton } from "features/tweet/views/TweetCardButton/CommentButton";
 import { UserBase } from "features/user/userTypes";
 import { formatDate } from "lib/functions/formatDate";
 import { deleteTweet } from "features/tweet/tweetApis";
